requirements.txt
flask
streamlit
tensorflow
numpy
pillow
requests
scikit-image